src/twitter/requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
amqp==2.6.1
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
blessings==1.7
boto3==1.17.84
botocore==1.20.112
cchardet==2.1.7
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.7
cssutils==2.3.0
dataclasses==0.6
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.3
django-3-jet==1.0.8
django-activity-stream==0.10.0
django-auto-prefetching==0.1.10
django-celery-beat==2.0.0
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-extensions==3.1.3
django-filter==2.4.0
django-health-check==3.16.4
django-inlinecss==0.3.0
django-model-utils==4.1.1
django-money==2.0
django-nose==1.4.7
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
django-summernote==0.8.11.6
django-timezone-field==4.2.1
django-unique-upload==0.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-yasg==1.20.0
easy-thumbnails==2.7.1
elasticsearch==7.14.0
factory-boy==3.2.0
fake-useragent==0.1.11
Faker==8.11.0
flake8==3.9.2
future==0.18.2
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
ipdb==0.13.8
ipython==7.24.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
multidict==5.1.0
newrelic==6.4.0.157
nose==1.3.7
nose-progressive==1.5.2
numpy==1.21.2
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.0
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
psycopg2==2.8.6
ptyprocess==0.7.0
py-moneyed==1.2
pycares==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.1.0
pynliner==0.8.0
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.2
python-socks==1.2.4
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.4.2
schedule==1.1.0
sentry-sdk==1.1.0
simplejson==3.17.3
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
-e git+https://github.com/twintproject/twint.git@e7c8a0c764f6879188e5c21e25fb6f1f856a7221#egg=twint
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
vine==1.3.0
wcwidth==0.2.5
yapf==0.31.0
yarl==1.6.3
